reset

+++ src/mbgl/webgpu/drawable.cpp
@@ -814,7 +814,7 @@ void Drawable::draw(PaintParameters& parameters) const {
 
         impl->pipelineState = shaderWebGPU.getRenderPipeline(renderable,
                                                              vertexLayouts.empty() ? nullptr : vertexLayouts.data(),
-                                                             static_cast<uint32_t>(vertexLayouts.size()),
+                                                             vertexLayouts.size(),
                                                              colorMode,
                                                              depthMode,
                                                              stencilMode,
@@ -935,9 +935,8 @@ void Drawable::draw(PaintParameters& parameters) const {
         const auto& segment = static_cast<DrawSegment&>(*seg_);
         const auto& mlSegment = segment.getSegment();
         if (mlSegment.indexLength > 0) {
-            const uint32_t instanceCount = instanceAttributes ? static_cast<uint32_t>(instanceAttributes->getMaxCount())
-                                                              : 1;
-            const uint32_t indexOffset = static_cast<uint32_t>(mlSegment.indexOffset);
+            const uint32_t instanceCount = instanceAttributes ? instanceAttributes->getMaxCount() : 1;
+            const uint32_t indexOffset = mlSegment.indexOffset;
             const int32_t baseVertex = static_cast<int32_t>(mlSegment.vertexOffset);
             const uint32_t baseInstance = 0;
 
@@ -954,11 +953,11 @@ void Drawable::draw(PaintParameters& parameters) const {
             }
 
             wgpuRenderPassEncoderDrawIndexed(renderPassEncoder,
-                                             static_cast<uint32_t>(mlSegment.indexLength), // indexCount
-                                             instanceCount,                                // instanceCount
-                                             indexOffset,                                  // firstIndex
-                                             baseVertex,                                   // baseVertex
-                                             baseInstance);                                // firstInstance
+                                             mlSegment.indexLength, // indexCount
+                                             instanceCount,         // instanceCount
+                                             indexOffset,           // firstIndex
+                                             baseVertex,            // baseVertex
+                                             baseInstance);         // firstInstance
 
             context.renderingStats().numDrawCalls++;
         }
